+++ packages/react-select/src/components/Group.js
@@ -1,7 +1,7 @@
 // @flow
 /** @jsx jsx */
 import { type Node, type ComponentType } from 'react';
-import { jsx } from '@emotion/core';
+import { jsx } from '@emotion/react';
 import { cleanCommonProps } from '../utils';
 
 import type { CommonProps } from '../types';
@@ -9,10 +9,12 @@ import type { CommonProps } from '../types';
 type ComponentProps = {
   /** The children to be rendered. */
   children: Node,
-  /** Component to wrap the label, recieves headingProps. */
+  /** Component to wrap the label, receives headingProps. */
   Heading: ComponentType<any>,
   /** Props to pass to Heading. */
   headingProps: any,
+  /** Props to be passed to the group element. */
+  innerProps: {},
   /** Label to be displayed in the heading component. */
   label: Node,
   /* The data of the group. */
@@ -33,6 +35,7 @@ const Group = (props: GroupProps) => {
     getStyles,
     Heading,
     headingProps,
+    innerProps,
     label,
     theme,
     selectProps,
@@ -41,6 +44,7 @@ const Group = (props: GroupProps) => {
     <div
       css={getStyles('group', props)}
       className={cx({ group: true }, className)}
+      {...innerProps}
     >
       <Heading
         {...headingProps}