+++ packages/react-select/src/components/Menu.js
@@ -7,7 +7,7 @@ import {
   type ElementRef,
   type Node,
 } from 'react';
-import { jsx } from '@emotion/core';
+import { jsx } from '@emotion/react';
 import { createPortal } from 'react-dom';
 
 import {
@@ -324,8 +324,8 @@ const Menu = (props: MenuProps) => {
     <div
       css={getStyles('menu', props)}
       className={cx({ menu: true }, className)}
-      {...innerProps}
       ref={innerRef}
+      {...innerProps}
     >
       {children}
     </div>
@@ -375,9 +375,9 @@ export const MenuList = (props: MenuListComponentProps) => {
     className,
     cx,
     getStyles,
-    isMulti,
-    innerRef,
     innerProps,
+    innerRef,
+    isMulti,
   } = props;
   return (
     <div
@@ -473,6 +473,7 @@ export type MenuPortalProps = CommonProps & {
   appendTo: HTMLElement,
   children: Node, // ideally Menu<MenuProps>
   controlElement: HTMLElement,
+  innerProps: {},
   menuPlacement: MenuPlacement,
   menuPosition: MenuPosition,
 };
@@ -509,7 +510,10 @@ export class MenuPortal extends Component<MenuPortalProps, MenuPortalState> {
     const {
       appendTo,
       children,
+      className,
       controlElement,
+      cx,
+      innerProps,
       menuPlacement,
       menuPosition: position,
       getStyles,
@@ -529,7 +533,18 @@ export class MenuPortal extends Component<MenuPortalProps, MenuPortalState> {
 
     // same wrapper element whether fixed or portalled
     const menuWrapper = (
-      <div css={getStyles('menuPortal', state)}>{children}</div>
+      <div
+        css={getStyles('menuPortal', state)}
+        className={cx(
+          {
+            'menu-portal': true,
+          },
+          className
+        )}
+        {...innerProps}
+      >
+        {children}
+      </div>
     );
 
     return (