+++ docs/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @react-select/docs
 
+## 3.0.1
+
+### Patch Changes
+
+- [a016c878](https://github.com/JedWatson/react-select/commit/a016c87821d9289ef9c317c0c397d64a0824ce16) [#4420](https://github.com/JedWatson/react-select/pull/4420) Thanks [@Methuselah96](https://github.com/Methuselah96)! - Bump dependency on @babel/runtime in order to fix compatibility issues with Webpack 5
+
+- [0dbf0438](https://github.com/JedWatson/react-select/commit/0dbf043864ce7a7fa7d822182b4f1770aad5b036) [#4387](https://github.com/JedWatson/react-select/pull/4387) Thanks [@ebonow](https://github.com/ebonow)! - Update example for MultiSelectSort to prevent dragging on multiValueRemove component #4387
+
+- Updated dependencies [f600d13f]:
+- Updated dependencies [b5f9b0c5]:
+- Updated dependencies [a016c878]:
+- Updated dependencies [19b76342]:
+- Updated dependencies [10b5f5a5]:
+  - react-select@undefined
+
+## 3.0.0
+
+### Major Changes
+
+- [26b6325c](https://github.com/JedWatson/react-select/commit/26b6325c95113591e568451bc2296f98318a8dd9) [#4283](https://github.com/JedWatson/react-select/pull/4283) Thanks [@majgaard](https://github.com/majgaard)! - Upgrades Emotion dependency to v11.0.0
+
+  BREAKING CHANGE: The NonceProvider component now requires a `cacheKey` prop that corresponds to the `key` for the Emotion cache.
+
+### Patch Changes
+
+- Updated dependencies [2d5496d5]:
+- Updated dependencies [02050675]:
+- Updated dependencies [26b6325c]:
+- Updated dependencies [b2488bb5]:
+  - react-select@undefined
+
 ## 2.4.6
 
 ### Patch Changes